Harmonizing and extending date validation in ServicePeriod. Dates must be between 1900 and 2100.
Added new function ValidateDate() which replaces _IsValidDate and which adds messages to the problems collection with the given context. ValidateStartDate(), ValidateEndDate() and ValidateDates() are now using this new function.

This change also affected the way the date_exceptions were stored. Up to now the dictionary contained the exceptions coming from calendar_dates.txt but their context got lost. This had the effect that the errors could not be related to the correct line for errors in calendar_dates.txt. My chosen solution is to save the context for each exception date in the date_exceptions using a tuple of the exception_type and the context

Added a convenience method PopColumnSpecificException() to test/util.py which allows to pop exceptions according to their type, column name and file name. Harmonized the other Pop* methods.

Review at http://codereview.appspot.com/2981041/