Fixed handling of values not in __dict__ for deprecated columns.
The __getattr__() function in GtfsObjectBase is a fall-back if a variable value is not found in __dict__.
If the variable is named after a valid field name it just returns None. Up to now it did only check in the _FIELD_NAMES list for valid field names. This change will also return None for fields in  _DEPRECATED_FIELD_NAMES.

Codereview at http://codereview.appspot.com/4097041/